+++ crates/config/src/lib.rs
@@ -19,6 +19,14 @@ pub fn load_workspace_config_template() -> &'static str {
     include_str!("../templates/workspace.yml")
 }
 
+pub fn load_workspace_node_config_template() -> &'static str {
+    include_str!("../templates/workspace_node.yml")
+}
+
+pub fn load_workspace_typescript_config_template() -> &'static str {
+    include_str!("../templates/workspace_typescript.yml")
+}
+
 pub fn load_global_project_config_template() -> &'static str {
     include_str!("../templates/global_project.yml")
 }
