+++ crates/config/templates/workspace.yml
@@ -16,80 +16,7 @@ projects:
   {{ id }}: '{{ source }}'
   {%- endfor %}
 
-# Configures Node.js within the toolchain. moon manages its own version of Node.js
-# instead of relying on a version found on the host machine. This ensures deterministic
-# and reproducible builds across any machine.
-node:
-  # The version to use. Must be a semantic version that includes major, minor, and patch.
-  # We suggest using the latest active LTS version: https://nodejs.org/en/about/releases
-  version: '{{ node_version }}'
-
-  # The package manager to use when managing dependencies.
-  # Accepts "npm" (default), "pnpm", or "yarn".
-  packageManager: '{{ package_manager }}'
-
-  {%- if package_manager_version != "inherit" %}
-
-  # The version of the package manager (above) to use.
-  {{ package_manager }}:
-    version: '{{ package_manager_version }}'
-  {%- endif %}
-
-  # Add `node.version` as a constraint in the root `package.json` `engines`.
-  addEnginesConstraint: true
-
-  # Use the `package.json` name as an alias for the respective moon project.
-  # aliasPackageNames: 'name-and-scope'
-
-  # Dedupe dependencies after the lockfile has changed.
-  dedupeOnLockfileChange: true
-
-  # Version format to use when syncing dependencies within the project's `package.json`.
-  # dependencyVersionFormat: 'workspace'
-
-  # Infer and automatically create moon tasks from `package.json` scripts, per project.
-  # BEWARE: Tasks and scripts are not 1:1 in functionality, so please refer to the documentation.
-  # inferTasksFromScripts: false
-
-  # Sync a project's `dependsOn` as dependencies within the project's `package.json`.
-  syncProjectWorkspaceDependencies: true
-
-  # Sync `node.version` to a 3rd-party version manager's config file.
-  # Accepts "nodenv" (.node-version), "nvm" (.nvmrc), or none.
-  {% if node_version_manager == "nvm" -%}
-  syncVersionManagerConfig: 'nvm'
-  {%- elif node_version_manager == "nodenv" -%}
-  syncVersionManagerConfig: 'nodenv'
-  {%- else -%}
-  # syncVersionManagerConfig: 'nvm'
-  {%- endif %}
-
-# Configures how moon integrates with TypeScript.
-typescript:
-  # When `syncProjectReferences` is enabled and a dependent project reference
-  # *does not* have a `tsconfig.json`, automatically create one.
-  createMissingConfig: true
-
-  # Name of `tsconfig.json` file in the project root.
-  # projectConfigFileName: 'tsconfig.json'
-
-  # Name of `tsconfig.json` file in the workspace root.
-  # rootConfigFileName: 'tsconfig.json'
-
-  # Name of the config file in the workspace root that defines shared compiler options
-  # for all project reference based config files.
-  # rootOptionsConfigFileName: 'tsconfig.options.json'
-
-  # Update the project's `tsconfig.json` to route the `outDir` to moon's cache.
-  # routeOutDirToCache: true
-
-  # Sync a project's `dependsOn` as TypeScript project references within the
-  # project's `tsconfig.json` and the workspace root `tsconfig.json`.
-  syncProjectReferences: true
-
-  # Sync a project's project references as `paths` aliases.
-  # syncProjectReferencesToPaths: true
-
+{% if vcs_manager != "git" && vcs_default_branch != "master" %}
 # Configures the version control system to utilize within the workspace. A VCS
 # is required for determining touched (added, modified, etc) files, calculating file hashes,
 # computing affected files, and much more.
@@ -102,3 +29,4 @@ vcs:
   # local branch against. For git, this is is typically "master" or "main",
   # and must include the remote prefix (before /). For svn, this should always be "trunk".
   defaultBranch: '{{ vcs_default_branch }}'
+{%- endif %}
